Add files via upload
The Iris dataset is a classic dataset in the field of machine learning and statistics. It contains measurements of iris flowers, specifically their sepal and petal dimensions. The dataset comprises 150 samples, with each sample belonging to one of three species: Setosa, Versicolor, or Virginica. Each sample includes four features: sepal length, sepal width, petal length, and petal width, all measured in centimeters. The dataset is often used for classification tasks, where the goal is to predict the species of an iris flower based on its measurements. It serves as a standard benchmark for testing various machine learning algorithms and techniques.